refactored code to fix namespace issue
