+++ article-search/src/appmodel/data_model.py
